feat: reinterpret `-q` or `--quite` to show raw output from cargo test
By default, `-q` is not passed, meaning raw output is forwarded as
cargo test is running.
If `-q` is passed in, don't display raw output, show test tree directly.

Note: this commit also makes `--color` default to `always` because not
`-q` is the default, and we want forwarding looks pretty in color.